l10n_et_base: convert to new style directory / file structure
