+++ l10n_et_base/models/__init__.py
@@ -21,4 +21,6 @@
 ##############################################################################
 
 # flake8: noqa
-from . import models
+from . import res_company
+from . import res_country
+from . import res_partner
