+++ l10n_et_base/views/res_company.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        
+        <record id="view_company_form" model="ir.ui.view">
+            <field name="name">res.company.form.inherit</field>
+            <field name="model">res.company</field>
+            <field name="inherit_id" ref="base.view_company_form"/>
+            <field name="arch" type="xml">
+                <data>
+                    <xpath expr="//field[@name='name']" position="after">
+                        <field name="ethiopic_name" placeholder="ስም በአማርኛ" />
+                    </xpath>
+                    <xpath expr="//field[@name='vat']" position="after">
+                    	<field name="vat_number"/>
+                    </xpath>
+                    <xpath expr="//field[@name='vat']" position="attributes">
+                    	<attribute name="string">TIN</attribute>
+                    </xpath>
+                    <xpath expr="//field[@name='company_registry']" position="attributes">
+                    	<attribute name="string">Registration No.</attribute>
+                    </xpath>
+                </data>
+            </field>
+        </record>
+
+    </data>
+</openerp>
